requirements.txt
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.0.1
billiard==4.2.0
boto3==1.35.19
botocore==1.35.19
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.114.2
h11==0.14.0
idna==3.10
jmespath==1.0.1
kombu==5.4.1
motor==3.5.1
passlib==1.7.4
prompt_toolkit==3.0.47
pyasn1==0.6.1
pydantic==2.9.1
pydantic_core==2.23.3
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
redis==5.0.8
requests==2.32.3
rsa==4.9
s3transfer==0.10.2
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
stripe==10.11.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.30.6
vine==5.1.0
wcwidth==0.2.13
